fix(defcon): widen BIGINT end-to-end + writer/DDL parity guard
Cascade fix surfaced when PR #210's `monotonically_increasing_id().cast("long")`
finally let the DEFCON pipeline re-execute against fresh data. Three latent
schema lies fired in sequence on the first re-trigger:

1. `valued_schema.action_player_id = StringType()` vs pandas int64 →
   `pyarrow.lib.ArrowTypeError: Expected a string or bytes dtype, got int64`
   in applyInPandas.
2. `valued_schema.defcon_value = DoubleType()` vs bronze DDL FLOAT →
   `[DELTA_FAILED_TO_MERGE_FIELDS] defcon_value` on the replaceWhere write.
3. `stg_defcon__results.cast(... as int)` on five BIGINT-bearing ID columns →
   `[CAST_OVERFLOW] LONG → INT` on the not_null staging tests.

Each had been hidden by SkipGuard for many cycles. Fix:

- defcon_lite_360.py + defcon_lite_tracking.py: action_player_id LongType,
  defcon_value FloatType in both credits_schema and valued_schema.
- stg_defcon__results.sql: 5 ID columns cast as bigint, not int
  (competition_id, season_id, defender_player_id, defender_team_id,
  action_player_id).
- _marts__models.yml + fct_defcon_actions/fct_defcon_pressure/fct_defensive_values:
  12 contract data_types int → bigint across the three defcon marts; matching
  cast(null as bigint) updates to the {% else %} disabled-state placeholders;
  test_marts_live_schema.py expected types updated to match.

Defense in depth (ADR-002 §4 confirming application):
- src/tests/test_defcon_schema_parity.py: parses _RESULTS_SCHEMA DDL string
  and asserts column-name + Spark-type equality with valued_schema for both
  defcon_lite_360 and defcon_lite_tracking, plus cross-module DDL equality.
  Catches all four of the above drift classes at CI time.

Live-invariant calibration (test_marts_kimball_contracts.py):
- Removed the two defender player_key threshold tests on fct_defensive_values
  and fct_defcon_actions. The previous 16% structural floor reported in
  Phase 0 Task 0.7 was an artifact of the broken INT cast truncating
  `monotonically_increasing_id()` LONG values into the real player_id range
  by accident; the post-fix structural floor is 0.06% (469 / 828k), which
  is correct — synthetic 360-anonymous defenders should not resolve to real
  dim_players entries. The test cannot meaningfully regression-check on
  noise.
- Raised team_key threshold 0.10 → 0.99 on the same two marts. PR #209's
  teammate-flag derivation now resolves at 100% (828634/828634) on dev_gold.

End-to-end verification on dev_gold (compute_defcon_lite run_id=163853960124728,
dbt_build run_id=559814695983073, synced refresh, live invariants):

  total rows                        828,634
  match_key set                       100.0%
  team_key set                        100.0%
  action_player_key set               100.0%
  player_key (defender, synthetic)      0.06%   (expected — not tested)
  fct_defcon_actions_synced              COMPLETE
  fct_defcon_pressure_synced             COMPLETE
  fct_defensive_values_synced            COMPLETE
  test_marts_kimball_contracts.py       29/29 PASS
  test_marts_live_schema.py             8/8 PASS
  test_idsse_is_progressive_coverage    2/2 PASS

Wheel: 0.3.15 → 0.3.18.

Documentation:
- ADR-002 §4 now lists DEFCON valued_schema as a confirming application of
  the writer/target schema drift guard pattern.
- ADR-011 staged-rollout table moves PR 5 / PR 6 from Planned → Shipped
  with PR numbers and followup chain.
- TODO.md last-updated → 2026-04-27 with the full followup sequence.

+++ src/tests/test_marts_live_schema.py
@@ -164,11 +164,15 @@ def test_fct_action_values_legacy_match_id_matches_new_key(conn: object) -> None
 
 _FCT_DEFENSIVE_VALUES_EXPECTED_COLS: dict[str, str] = {
     "defensive_value_id": "string",
-    "player_id": "int",
+    # bigint, not int — bronze.defcon_results stores BIGINT IDs because the
+    # 360-synthetic defender_player_id from monotonically_increasing_id()
+    # exceeds INT range under multi-partition serverless. Widening propagated
+    # end-to-end 2026-04-27 to fix CAST_OVERFLOW on staging not_null tests.
+    "player_id": "bigint",
     "match_id": "string",
-    "competition_id": "int",
-    "season_id": "int",
-    "team_id": "int",
+    "competition_id": "bigint",
+    "season_id": "bigint",
+    "team_id": "bigint",
     "data_source": "string",
     # PR 6 Kimball surrogates
     "match_key": "bigint",
@@ -194,13 +198,14 @@ _FCT_DEFCON_ACTIONS_EXPECTED_COLS: dict[str, str] = {
     "defcon_action_id": "string",
     "event_id": "string",
     "match_id": "string",
-    "competition_id": "int",
-    "season_id": "int",
-    "player_id": "int",
-    "team_id": "int",
+    # bigint — see _FCT_DEFENSIVE_VALUES_EXPECTED_COLS rationale.
+    "competition_id": "bigint",
+    "season_id": "bigint",
+    "player_id": "bigint",
+    "team_id": "bigint",
     "defender_x": "double",
     "defender_y": "double",
-    "action_player_id": "int",
+    "action_player_id": "bigint",
     "action_type": "string",
     "action_x": "double",
     "action_y": "double",
@@ -221,10 +226,11 @@ _FCT_DEFCON_ACTIONS_EXPECTED_COLS: dict[str, str] = {
 
 _FCT_DEFCON_PRESSURE_EXPECTED_COLS: dict[str, str] = {
     "pressure_id": "string",
-    "player_id": "int",
+    # bigint — see _FCT_DEFENSIVE_VALUES_EXPECTED_COLS rationale.
+    "player_id": "bigint",
     "match_id": "string",
-    "competition_id": "int",
-    "season_id": "int",
+    "competition_id": "bigint",
+    "season_id": "bigint",
     "data_source": "string",
     # PR 6 Kimball surrogates
     "match_key": "bigint",
